Refactor stack and opcodes into their own modules
