+++ coco-core/src/lib.rs
@@ -3,6 +3,11 @@
 
 use core::fmt;
 
+mod stack;
+use stack::Stack;
+
+mod opcodes;
+
 /// The trait to implement for COCO virtual machines.
 pub trait Machine {
     fn deo(&mut self, cpu: &mut Cpu, target: u8) -> bool;
@@ -12,41 +17,6 @@ pub trait Machine {
 /// The trait to implement a COCO device's ports
 pub trait Ports {
     const BASE: u8;
-}
-
-#[derive(Debug)]
-struct Stack {
-    data: [u8; 0x100],
-    index: u8,
-}
-
-impl Stack {
-    pub fn new() -> Self {
-        Self {
-            data: [0_u8; 0x100],
-            index: 0,
-        }
-    }
-
-    fn byte_at(&self, i: u8) -> u8 {
-        return self.data[i as usize];
-    }
-}
-
-impl fmt::Display for Stack {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        for i in self.index.saturating_sub(8)..self.index {
-            write!(f, "{:02x}", self.byte_at(i))?;
-            if i < self.index - 1 {
-                write!(f, " ")?;
-            }
-        }
-        Ok(())
-    }
-}
-
-mod opcodes {
-    pub const BRK: u8 = 0x00;
 }
 
 /// COCO-8 CPU.
