+++ coco-core/src/opcodes.rs
@@ -1,0 +1,1 @@
+pub const BRK: u8 = 0x00;
